% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{custom_upset_plot}
\alias{custom_upset_plot}
\title{custom_upset_plot}
\usage{
custom_upset_plot(data, query_non_acute, title)
}
\arguments{
\item{data}{a data object, preferably a data frame object}

\item{query_non_acute}{a concatenated String vector of names to be displayed,
excluding the non acute wounds}

\item{title}{a title for the ggplot to be provided}
}
\value{
upset_plot a ggplot2 object, which contains an upset plot
}
\description{
custom_upset_plot
}
