% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{exclude_rowname}
\alias{exclude_rowname}
\title{exclude_rowname}
\usage{
exclude_rowname(data, str_idx)
}
\arguments{
\item{data}{an data object - ideally a data frame}

\item{str_idx}{the indice to be excluded, provided as a string}
}
\value{
data returns the given dataset without the provided indices
}
\description{
exclude_rowname
}
