% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{export_ggplot}
\alias{export_ggplot}
\title{export_ggplot}
\usage{
export_ggplot(file_name, plot_name, width, height)
}
\arguments{
\item{file_name}{file name of the to be saved object - needs to be a string}

\item{plot_name}{ggplot2 object}

\item{width}{needs to be a numeric, e.g., 85 * (14/5)}

\item{height}{needs to be a numeric e.g., 85 * (14/5)}
}
\description{
uses the ggsave function to save the passed on ggplot2 object as a png to
the passed path given by the file_name parameter
}
