% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_upset_queries}
\alias{generate_upset_queries}
\title{generate_upset_queries}
\usage{
generate_upset_queries(names)
}
\arguments{
\item{names}{a named list object}
}
\value{
queries - a list of upset plot queries
}
\description{
iterates through the given named list and creates queries for all the individual
names created for an upset plot.
}
