% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{get_names_and_max_idx}
\alias{get_names_and_max_idx}
\title{get_names_and_max_idx}
\usage{
get_names_and_max_idx(data)
}
\arguments{
\item{data}{an data object, preferably a data frame}
}
\value{
name_frequencies returns a list object with the unique colnames and
there respective freqency of occurrence
}
\description{
get_names_and_max_idx
}
