% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{merge_binary_res}
\alias{merge_binary_res}
\title{merge_binary_res}
\usage{
merge_binary_res(data, ent_name, max_idx)
}
\arguments{
\item{data}{an data object, preferably a data frame}

\item{ent_name}{the name of the entity to be merged, needs to be a string}

\item{max_idx}{the max index this entity has in the dataframe}
}
\value{
data returns the provided dataset, with the modifications made
}
\description{
merge_binary_res
}
